--- src/app/components/elements/Modal/HowToList.ts
@@ -1,34 +1,0 @@
-const howToList = [
-  {
-    number: '01',
-    title: '目標(Reach)を入力します。',
-    description:'まずは成し遂げたい目標を登録します。\nなるべく具体的な目標を登録することがおすすめです。\n具体的な目標を決めることで、達成するために必要な要素を明確にとらえることができます。',
-    image: ''
-  },
-  {
-    number: '02',
-    title: '達成するために必要な「8つの要素(Skill)」を登録していきます。',
-    description:'8つの要素を洗い出すことで、バランス、思考の整理、行動計画の具体化、柔軟な思考が得られやすいと言われています。目標の達成の種類にもよりますが、要素として「心・技・体・生活」をバランスよく配分することが大切です。\n8つ以上の要素を登録することができますが、優先順位をつけて取捨選択することが推奨されています。',
-    image: ''
-  },
-  {
-    number: '03',
-    title: '必要な要素を実現するために、要素ごとに[８つの行動目標(Action)]を登録していきます。',
-    description:'「8つの行動目標」には、可能な限り具体的な数字や期日などを定めましょう。\n記述を決めた行動とルーティン行動の2種類を設定することが推奨されています。',
-    image: ''
-  },
-  {
-    number: '04',
-    title: '進捗の確認',
-    description:'Actionが完了したら、チェックマークをつけて進捗度を可視化することもできます。',
-    image: ''
-  },
-  {
-    number: '05',
-    title: '振り返り',
-    description:'目標達成に向けた行動の振り返りを記入して、必要に応じて要素の修正や行動を修正します。',
-    image: ''
-  }
-]
-
-export default howToList;
