+++ src/app/components/elements/Modal/HowToModal/HowTo.module.scss
@@ -1,0 +1,123 @@
+@use "@/styles/_variables.scss" as *;
+
+.container {
+  width: 100vw;
+  height: 100vh;
+  position: fixed;
+  top: 0;
+  left: 0;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  z-index: 99999;
+
+  & .backround {
+    display: block;
+    width: 100vw;
+    height: 100vh;
+    position: fixed;
+    top: 0;
+    left: 0;
+    background-color: rgb(110, 110, 110);
+    opacity: .9;
+  }
+
+  & .content {
+    width: 100%;
+    max-width: 750px;
+    height: 100%;
+    padding: 2rem 1rem;
+    position: relative;
+    background-color: $baseColor;
+    border-radius: 10px;
+    display: grid;
+    grid-template-rows: 1fr 1fr 5fr;
+    gap: 1rem;
+    transform: translateY(-3rem);
+    overflow: scroll;
+
+    @include mq(lg) {
+      margin: 0 auto;
+      padding: 3rem;
+      transform: translateY(0);
+    }
+
+
+    & .modal_title {
+      font: bold 1.5rem var(--font-jp);
+      color: $red;
+
+      & .step {
+        font: bold 2.5rem var(--font-en);
+        color: white;
+      }
+    }
+
+    & .description {
+      font: normal 1rem var(--font-jp);
+      white-space: pre-line;
+    }
+
+    & .image_container{
+      width: 100%;
+      height: 100%;
+      position: relative;
+    }
+  }
+
+  & .close_button {
+    width: 2rem;
+    aspect-ratio: 1/1;
+    position: absolute;
+    top: 2rem;
+    right: 1rem;
+
+    & .close_line1,
+    .close_line2 {
+      display: block;
+      width: 100%;
+      height: 1px;
+      background-color: white;
+    }
+
+    & .close_line1 {
+      transform: translateY(3px) rotate(180deg);
+      transition: all .2s .1s $ease;
+    }
+
+    & .close_line2 {
+      transform: translateY(-3px);
+      transition: all .2s $ease;
+    }
+
+    &:hover {
+      & .close_line1 {
+        transform: rotate(15deg);
+      }
+
+      & .close_line2 {
+        transform: translateY(-1px) rotate(345deg);
+      }
+    }
+  }
+}
+
+.swiper_container {
+  width: 100vw;
+  height: 100vh;
+
+  & .swiper_slide {
+    height: 65%;
+    justify-self: center;
+    align-self: center;
+
+    @include mq(lg){
+      height: 80%;
+    }
+  }
+}
+
+.swiper-button-prev{
+  color: rebeccapurple;
+}
