+++ src/app/features/login/Index.tsx
@@ -1,13 +1,15 @@
 "use client"
 
-import HowToModal from "@/components/elements/Modal/HowToModal";
+
 import { useIsHowToModal } from "@/components/utils/IsHowToModalProvider";
 import ModalContainer from "@/components/utils/ModalContainer";
+import HowToModal from "@/components/elements/Modal/HowToModal/HowToModal";
 import Description from "./Description";
+import Start from "./Start";
 import Issue from "./Issue";
 import Features from "./Features";
+
 import styles from "./styles/Login.module.scss"
-import Start from "./Start";
 
 export default function LoginIndex() {
